Use an ABC for executors
