+++ app/grandchallenge/workstations/models.py
@@ -423,8 +423,6 @@ class Session(UUIDModel):
             job_id=f"{self._meta.app_label}-{self._meta.model_name}-{self.pk}",
             exec_image_repo_tag=self.workstation_image.original_repo_tag,
             memory_limit=settings.COMPONENTS_MEMORY_LIMIT,
-            requires_gpu=False,
-            time_limit=settings.WORKSTATIONS_SESSION_DURATION_LIMIT,
         )
 
     @property
